include update tests

+++ e2eTests/e2eTestHelper.psm1
@@ -191,9 +191,10 @@ function invoke-git {
     if ($lastexitcode) { throw "git $command error" }
 }
 
-function CreateAndCloneRepository {
+function CreateRepository {
     Param(
         [string] $template,
+        [string] $templatePath,
         [switch] $private,
         [string] $branch = "main"
     )
@@ -212,25 +213,32 @@ function CreateAndCloneRepository {
     }
     Set-Location '*'
 
-    $templateUrl = "$template/archive/refs/heads/$branch.zip"
-    $zipFileName = Join-Path $tempPath "$([GUID]::NewGuid().ToString()).zip"
-    [System.Net.WebClient]::new().DownloadFile($templateUrl, $zipFileName)
-    
-    $tempRepoPath = Join-Path $tempPath ([GUID]::NewGuid().ToString())
-    Expand-Archive -Path $zipFileName -DestinationPath $tempRepoPath
-    Copy-Item (Join-Path (Get-Item "$tempRepoPath\*").FullName '*') -Destination . -Recurse -Force
-    Remove-Item -Path $tempRepoPath -Force -Recurse
-    Remove-Item -Path $zipFileName -Force
+    if ($template) {
+        $templateUrl = "$template/archive/refs/heads/$branch.zip"
+        $zipFileName = Join-Path $tempPath "$([GUID]::NewGuid().ToString()).zip"
+        [System.Net.WebClient]::new().DownloadFile($templateUrl, $zipFileName)
+        
+        $tempRepoPath = Join-Path $tempPath ([GUID]::NewGuid().ToString())
+        Expand-Archive -Path $zipFileName -DestinationPath $tempRepoPath
+        Copy-Item (Join-Path (Get-Item "$tempRepoPath\*").FullName '*') -Destination . -Recurse -Force
+        Remove-Item -Path $tempRepoPath -Force -Recurse
+        Remove-Item -Path $zipFileName -Force
+    }
+    if ($templatePath) {
+        Write-Host "$(Join-Path $templatePath '*')"
+
+        Copy-Item (Join-Path $templatePath '*') -Destination . -Recurse -Force
+    }
 
     invoke-git add *
-    invoke-git commit -m 'init'
+    invoke-git commit --allow-empty -m 'init'
     invoke-git branch -M $branch
     if ($githubOwner -and $token) {
         invoke-git remote set-url origin "https://$($githubOwner):$token@github.com/$repository.git"
     }
     invoke-git push --set-upstream origin $branch
 
-    Start-Sleep -seconds $delay
+   Start-Sleep -seconds $delay
 }
 
 function Pull {
@@ -248,7 +256,7 @@ function CommitAndPush {
     )
 
     invoke-git add *
-    invoke-git commit -m "$commitMessage"
+    invoke-git commit --allow-empty -m "$commitMessage"
     invoke-git push $serverUrl | Out-Host
 }
